// src/types/models/televiewerDisc.ts
export interface TeleviewerDisc {
  rpId: string;
  dip: number;
  dipDirection: number;
  pX: number;
  pY: number;
  pZ: number;
  nX: number;
  nY: number;
  nZ: number;
  holeNumber: number;
  typeOfHole: string;
  imageOrMeshOfFractureInterpolation: string;
  explanation: string;
  zAdjust: number;
}